+++ src/app/(public)/layout.tsx
@@ -8,8 +8,8 @@ import { UserAuth } from '@/app/_components/user-auth'
 import { Header } from '@/app/_components/header'
 import { NavMenu } from '@/app/_components/nav-menu'
 import { Footer } from '@/app/_components/footer'
-import { FeedbackDialog } from '@/app/_components/feedback-dialog'
 import { StarsBackground } from '@/ui/stars-background'
+import { LinksMenu } from '@/app/_components/links-menu'
 
 const Layout = async ({ children }: { children: React.ReactNode }) => {
   const user = await api.auth.getUser()
@@ -21,20 +21,22 @@ const Layout = async ({ children }: { children: React.ReactNode }) => {
       </Suspense>
 
       <Header>
-        <div className="flex items-end">
+        <LinksMenu className="sm:hidden" />
+
+        <div className="hidden w-32 items-end justify-start sm:flex">
           <Logo />
           <span className="ml-1 hidden rounded-md bg-white/10 px-1.5 py-0.5 text-xs font-medium text-white/80 opacity-50 sm:block">
             v{VERSION}
           </span>
         </div>
 
-        <NavMenu />
-        <UserAuth user={user} />
+        <NavMenu className="hidden sm:block" />
+        <div className="flex w-32 justify-end">
+          <UserAuth user={user} />
+        </div>
       </Header>
 
       {children}
-
-      {user && <FeedbackDialog email={user.email as string} />}
       <Footer />
     </div>
   )