+++ src/server/routers/groups.ts
@@ -1,27 +1,33 @@
-import { z } from 'zod'
+import { groups } from '@prisma/client'
+import { z } from 'zod/v4'
 
 import { createTRPCRouter, publicProcedure } from '@/server/trpc'
 import { db } from '@/prisma/db'
-import { MAX_LEETCODERS } from '@/data/constants'
+import { MAX_LEETCODERS, REDIS_KEYS } from '@/data/constants'
+import { redis } from '@/config/redis'
+import type { GetAllGroupsInfoType, GroupTableDataType } from '@/trpc/groups.type'
+import { GetAllAvailableGroupsType } from '@/types/groups.type'
+import { logger } from '@/utils/logger'
 
 export const groupsRouter = createTRPCRouter({
-  getAllGroupsNo: publicProcedure.query(() => {
-    return db.groups.findMany({
-      select: { group_no: true },
-    })
-  }),
-  getGroupByNo: publicProcedure
-    .input(z.object({ group_no: z.string().transform((val) => Number(val)) }))
-    .query(({ input }) => {
-      return db.groups.findUnique({
-        where: { group_no: input.group_no },
-        select: { group_no: true },
-      })
-    }),
+  /**
+   * Get group table data for /group/[groupId] page
+   * - revalidates every 24 hours
+   * - on each leetcoder submission, the cache is invalidated
+   * @returns {GroupTableDataType}
+   */
   getGroupTableData: publicProcedure
     .input(z.object({ group_no: z.string().transform((val) => Number(val)) }))
-    .query(({ input }) => {
-      return db.groups.findUnique({
+    .query(async ({ input }): Promise<GroupTableDataType> => {
+      const cachedGroupData = (await redis.get(
+        REDIS_KEYS.GROUP_DATA(input.group_no.toString())
+      )) as GroupTableDataType | null
+      if (cachedGroupData) {
+        logger.debug(`[Cache] Using cached group data for group ${input.group_no}`)
+        return cachedGroupData
+      }
+
+      const groupData = await db.groups.findUnique({
         where: {
           group_no: input.group_no,
         },
@@ -40,11 +46,38 @@ export const groupsRouter = createTRPCRouter({
           },
         },
       })
+
+      await redis.set(REDIS_KEYS.GROUP_DATA(input.group_no.toString()), groupData, { ex: 86400 }) // cache for one day
+      return groupData as GroupTableDataType
     }),
+  /**
+   * Get all groups for /dashboard page
+   * revalidates every 7 days
+   * @returns {groups[]}
+   */
   getAllGroups: publicProcedure.query(async () => {
-    return db.groups.findMany()
+    const cachedGroups = (await redis.get(REDIS_KEYS.ALL_GROUPS)) as groups[] | null
+    if (cachedGroups) {
+      logger.debug(`[Cache] Using cached all groups data`)
+      return cachedGroups
+    }
+
+    const groups = await db.groups.findMany()
+
+    await redis.set(REDIS_KEYS.ALL_GROUPS, groups, { ex: 604800 }) // cache for 7 days
+    return groups
   }),
+  /**
+   * Get all available groups for /profile page
+   * revalidates every 24 hours
+   */
   getAllAvailableGroups: publicProcedure.query(async () => {
+    const cachedGroups = (await redis.get(REDIS_KEYS.AVAILABLE_GROUPS)) as GetAllAvailableGroupsType[] | null
+    if (cachedGroups) {
+      logger.debug(`[Cache] Using cached available groups data`)
+      return cachedGroups
+    }
+
     const groups = await db.groups.findMany({
       include: {
         leetcoders: {
@@ -63,10 +96,25 @@ export const groupsRouter = createTRPCRouter({
       },
     })
 
-    return groups.filter((group) => group.leetcoders.length < MAX_LEETCODERS)
+    const availableGroups = groups.filter((group) => group.leetcoders.length < MAX_LEETCODERS)
+    await redis.set(REDIS_KEYS.AVAILABLE_GROUPS, availableGroups, { ex: 86400 }) // cache for 24 hours
+    return availableGroups as GetAllAvailableGroupsType[]
   }),
-  getAllGroupsInfo: publicProcedure.query(() => {
-    return db.groups.findMany({
+  /**
+   * Get all groups info for /groups page
+   * revalidates every 24 hours
+   * - on requestToJoinGroup, the cache is invalidated
+   * - on each leetcoder change group request, the cache is invalidated
+   * @returns {GetAllGroupsInfoType[]}
+   */
+  getAllGroupsInfo: publicProcedure.query(async () => {
+    const cachedGroupsInfo = (await redis.get(REDIS_KEYS.ALL_GROUPS_INFO)) as GetAllGroupsInfoType[] | null
+    if (cachedGroupsInfo) {
+      logger.debug(`[Cache] Using cached all groups info data`)
+      return cachedGroupsInfo
+    }
+
+    const groupsInfo = await db.groups.findMany({
       orderBy: {
         group_no: 'asc',
       },
@@ -90,21 +138,8 @@ export const groupsRouter = createTRPCRouter({
         },
       },
     })
-  }),
-  getGroupLeetcodersCount: publicProcedure.query(async () => {
-    try {
-      const count = await db.leetcoders.count({
-        where: {
-          status: {
-            in: ['APPROVED', 'PENDING'],
-          },
-        },
-      })
 
-      return count
-    } catch (error) {
-      console.error('Error fetching approved leetcoders count:', error)
-      throw new Error('Failed to fetch approved leetcoders count')
-    }
+    await redis.set(REDIS_KEYS.ALL_GROUPS_INFO, groupsInfo, { ex: 86400 }) // cache for one day
+    return groupsInfo as GetAllGroupsInfoType[]
   }),
 })