+++ src/types/groups.type.ts
@@ -1,0 +1,5 @@
+import type { groups, leetcoders } from '@prisma/client'
+
+export type GetAllAvailableGroupsType = groups & {
+  leetcoders: Pick<leetcoders, 'id'>[]
+}