+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # Changelog
 
+## [2.5.31]
+update localStorage key to fix sumissions issue, reload after success submission
+
+## [2.5.30]
+update navmenu in small screens
+
+## [2.5.29]
+remove feedback dialog
+
+## [2.5.28]
+updated the automated backup frequency to run every 6 hours
+
+## [2.5.27]
+close nav menu when a link is clicked
+
+## [2.5.26]
+add links menu in small screens
+
+## [2.5.25]
+Add Prisma Optimize extension to client
+
+## [2.5.24]
+add techs resources
+
+## [2.4.24]
+fix roadmap in small screens
+
+## [2.4.23]
+roadmap is a protected route
+
+## [2.4.21]
+cache all data for better performance
+
+## [2.3.20]
+hide the request button when logged in user in the group
+
 ## [2.3.19]
 fix group change func